Fix comparisons of integers of different signs

+++ src/data/corpus.cpp
@@ -51,21 +51,21 @@ Corpus::Corpus(Ptr<Config> options, bool translate)
   if(!translate) {
     std::vector<Vocab> vocabs;
     if(vocabPaths.empty()) {
-      for(int i = 0; i < paths_.size(); ++i) {
+      for(size_t i = 0; i < paths_.size(); ++i) {
         Ptr<Vocab> vocab = New<Vocab>();
         vocab->loadOrCreate("", paths_[i], maxVocabs[i]);
         options_->get()["vocabs"].push_back(paths_[i] + ".yml");
         vocabs_.emplace_back(vocab);
       }
     } else {
-      for(int i = 0; i < vocabPaths.size(); ++i) {
+      for(size_t i = 0; i < vocabPaths.size(); ++i) {
         Ptr<Vocab> vocab = New<Vocab>();
         vocab->loadOrCreate(vocabPaths[i], paths_[i], maxVocabs[i]);
         vocabs_.emplace_back(vocab);
       }
     }
   } else {
-    for(int i = 0; i < vocabPaths.size() - 1; ++i) {
+    for(size_t i = 0; i < vocabPaths.size() - 1; ++i) {
       Ptr<Vocab> vocab = New<Vocab>();
       vocab->loadOrCreate(vocabPaths[i], paths_[i], maxVocabs[i]);
       vocabs_.emplace_back(vocab);
@@ -106,7 +106,7 @@ SentenceTuple Corpus::next() {
       curId = ids_[pos_++];
 
     SentenceTuple tup(curId);
-    for(int i = 0; i < files_.size(); ++i) {
+    for(size_t i = 0; i < files_.size(); ++i) {
       std::string line;
       if(std::getline((std::istream&)*files_[i], line)) {
         Words words = (*vocabs_[i])(line);
@@ -157,7 +157,7 @@ void Corpus::shuffleFiles(const std::vector<std::string>& paths) {
   bool cont = true;
   while(cont) {
     std::vector<std::string> lines(files_.size());
-    for(int i = 0; i < files_.size(); ++i) {
+    for(size_t i = 0; i < files_.size(); ++i) {
       cont = cont && std::getline((std::istream&)*files_[i], lines[i]);
     }
     if(cont)
@@ -172,7 +172,7 @@ void Corpus::shuffleFiles(const std::vector<std::string>& paths) {
   tempFiles_.clear();
 
   std::vector<UPtr<OutputFileStream>> outs;
-  for(int i = 0; i < files_.size(); ++i) {
+  for(size_t i = 0; i < files_.size(); ++i) {
     tempFiles_.emplace_back(
         new TemporaryFile(options_->get<std::string>("tempdir")));
     outs.emplace_back(new OutputFileStream(*tempFiles_[i]));
@@ -187,7 +187,7 @@ void Corpus::shuffleFiles(const std::vector<std::string>& paths) {
   }
 
   files_.clear();
-  for(int i = 0; i < outs.size(); ++i) {
+  for(size_t i = 0; i < outs.size(); ++i) {
     files_.emplace_back(new InputFileStream(*tempFiles_[i]));
   }
